Checksum code moved to own files.  Generic implementation now allows for other methods of checksum.
